Login 2

+++ murrikulum/src/main/java/app/backend/repository/EnpresaRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.repository.CrudRepository;
 import app.backend.model.Enpresa;
 
 public interface EnpresaRepository extends CrudRepository<Enpresa, Long> {
-    Optional<Enpresa> findByUserId(Long userId);
+    Optional<Enpresa> findByUserId(Integer userId);
 }
